examples/csv_agent/requirements.txt
lihil-mcp
pandas>=2.0.0
numpy>=1.24.0
uvicorn>=0.20.0